scraper/requirements.txt
feedparser==6.0.11
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
